+++ lib/main.dart
@@ -47,7 +47,7 @@ Future<void> $main({
       options.environment = getSentryEnvironment();
       options.tracesSampleRate = 0.5;
       options.profilesSampleRate = 0.5;
-      options.experimental.replay.onErrorSampleRate = 1.0;
+      options.replay.onErrorSampleRate = 1.0;
       options.beforeSend = (event, hint) {
         // Filter out network related errors to prevent noise in Sentry
         if (event.throwable is SocketException ||
@@ -120,7 +120,7 @@ Future<void> $main({
       TimeMeasurements.mainFunction.stop();
       debugPrint('Main function done after ${TimeMeasurements.mainFunction.elapsedMilliseconds}ms');
 
-      runApp(maybeWrappedApp);
+      runApp(SentryWidget(child: maybeWrappedApp));
     },
   );
 }