+++ lib/screens/episode.dart
@@ -12,12 +12,14 @@ import 'package:brunstadtv_app/components/episode/episode_collection.dart';
 import 'package:brunstadtv_app/components/episode/episode_info.dart';
 import 'package:brunstadtv_app/components/episode/episode_related.dart';
 import 'package:brunstadtv_app/components/episode/episode_season.dart';
+import 'package:brunstadtv_app/components/player/audio_only_player.dart';
 import 'package:brunstadtv_app/components/player/player_poster.dart';
 import 'package:brunstadtv_app/components/misc/parental_gate.dart';
 import 'package:brunstadtv_app/components/status/error_generic.dart';
 import 'package:brunstadtv_app/components/status/loading_indicator.dart';
 import 'package:brunstadtv_app/components/episode/episode_share_sheet.dart';
 import 'package:brunstadtv_app/components/video/center_extra_slot.dart';
+import 'package:brunstadtv_app/providers/audio_only_provider.dart';
 import 'package:brunstadtv_app/providers/lesson_progress_provider.dart';
 import 'package:brunstadtv_app/router/router.gr.dart';
 import 'package:collection/collection.dart';
@@ -48,7 +50,7 @@ class EpisodeScreen extends _EpisodeScreenImplementation {
   const EpisodeScreen({
     super.key,
     @PathParam() required super.episodeId,
-    @QueryParam() super.autoplay,
+    @QueryParam() super.autoplay = true,
     @QueryParam('t') super.queryParamStartPositionSeconds,
     @QueryParam('hide_bottom_section') super.hideBottomSection,
     @QueryParam('collectionId') super.collectionId,
@@ -60,7 +62,7 @@ class CollectionEpisodeScreen extends _EpisodeScreenImplementation {
   const CollectionEpisodeScreen({
     super.key,
     @PathParam() required super.episodeId,
-    @QueryParam() super.autoplay,
+    @QueryParam() super.autoplay = true,
     @QueryParam('t') super.queryParamStartPositionSeconds,
     @QueryParam('hide_bottom_section') super.hideBottomSection,
     @PathParam('collectionId') super.collectionId,
@@ -298,7 +300,7 @@ class _EpisodeDisplay extends HookConsumerWidget {
           controlsConfig: defaultViewConfig.controlsConfig.copyWith(
             extraSettingsBuilder: (context) {
               return [
-                if (languages.whereNotNull().isNotEmpty)
+                if (languages.nonNulls.isNotEmpty)
                   VideoLanguageSettings(
                     episode: episode,
                     languages: languages.toList(),
@@ -387,6 +389,14 @@ class _EpisodeDisplay extends HookConsumerWidget {
 
     final playerError = useListenableSelector(viewController.playerController, () => viewController.playerController.value.error);
 
+    final forceAudioOnly = useState(false);
+    useEffect(() {
+      final isAudio = episode.streams.firstWhereOrNull((stream) => stream.primaryMediaType == Enum$PrimaryMediaType.audio) != null;
+      if (isAudio) {
+        forceAudioOnly.value = true;
+      }
+    }, [episode.streams]);
+
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
@@ -419,7 +429,12 @@ class _EpisodeDisplay extends HookConsumerWidget {
                   MediaQuery.removePadding(
                     context: context,
                     removeBottom: true,
-                    child: BccmPlayerView.withViewController(viewController),
+                    child: AnimatedSize(
+                        duration: const Duration(milliseconds: 800),
+                        curve: Curves.easeOutExpo,
+                        child: ref.watch(audioOnlyProvider) || forceAudioOnly.value
+                            ? AudioOnlyPlayer(viewController: viewController, disableToggle: forceAudioOnly.value)
+                            : BccmPlayerView.withViewController(viewController)),
                   ),
                 AnimatedSize(
                   duration: const Duration(milliseconds: 800),